+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [4.11.0](https://github.com/googleapis/nodejs-bigquery-storage/compare/v4.10.1...v4.11.0) (2025-01-09)
+
+
+### Features
+
+* Emit connection event for a StreamConnection ([#533](https://github.com/googleapis/nodejs-bigquery-storage/issues/533)) ([6688ea3](https://github.com/googleapis/nodejs-bigquery-storage/commit/6688ea3e87e7bd992802d643daa1f760c0c4dd92))
+
+
+### Bug Fixes
+
+* Required table fields with value expression should be proto optional ([#532](https://github.com/googleapis/nodejs-bigquery-storage/issues/532)) ([f125792](https://github.com/googleapis/nodejs-bigquery-storage/commit/f1257924be9bb257ea9a11448322f8aa711ad47e))
+
+## [4.10.1](https://github.com/googleapis/nodejs-bigquery-storage/compare/v4.10.0...v4.10.1) (2024-10-23)
+
+
+### Bug Fixes
+
+* Arrow parsing for nested and nullable fields ([#519](https://github.com/googleapis/nodejs-bigquery-storage/issues/519)) ([fc052fb](https://github.com/googleapis/nodejs-bigquery-storage/commit/fc052fb8746f0f87e028f009a6eccc78ff8a1bc4))
+
 ## [4.10.0](https://github.com/googleapis/nodejs-bigquery-storage/compare/v4.9.0...v4.10.0) (2024-09-23)